corrigiendo versiones de los proyectos

+++ netbeans/bancoPersistencia/src/main/java/banco/bancopersistencia/conexion/controladores/ControlPersistencia.java
@@ -15,6 +15,7 @@ import banco.bancopersistencia.daos.ICuentaDAO;
 import banco.bancopersistencia.dtos.ClienteDTO;
 import banco.bancopersistencia.dtos.CuentaDTO;
 import banco.bancopersistencia.excepciones.PersistenciaException;
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -78,7 +79,7 @@ public class ControlPersistencia implements IControlPersistencia{
     }
 
     @Override
-    public void actualizarSaldoCuenta(String num_cuenta, float saldo) throws PersistenciaException {
+    public void actualizarSaldoCuenta(String num_cuenta, BigDecimal saldo) throws PersistenciaException {
         this.cuentaDAO.actualizarSaldoCuenta(num_cuenta, saldo);
     }
